Add Services directory, rename dev_request to request_service and move file to new directory
